requirements.txt
######-Requirements-######

#Pwntools: https://docs.pwntools.com/en/stable/#
pwntools
#Colorama: https://pypi.org/project/colorama/#
colorama
